requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.4
django-cors-headers==4.7.0
django-restframework==0.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
idna==3.10
mercadopago==2.3.0
pillow==11.3.0
PyJWT==2.10.1
requests==2.32.5
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.7.4.3
typing_extensions==4.14.1
urllib3==2.5.0
